Redesign packets recording
Now you have to click the record button to start recording packets. This will help with performance and also save some bandtwitdh between browser and the server.

How this works is basically that server broadcasts packets to all websocket clients all the time. When you press record a new websocket connection is being created and frontend receives the broadcasted packets. When you stop recording the websocketconnection is ended and client no longer receives the packets. Then user is able to see the packets in the client.